登陆注册功能已完成，对于登录和注册功能在app-server模块的测试包下创建了测试类进行了单元测试，同时完成登录和注册的接口测试。user数据库表语句有更新，将密码数据类型更改查varchar(100)，u_id添加自增。

+++ SMS-model/src/main/java/supermarket/manage/system/model/entity/AuthUserEntity.java
@@ -12,14 +12,14 @@ import java.util.Collection;
 public class AuthUserEntity extends User {
 
     /** 用户id */
-    private Long id;
+    private Integer id;
 
-    public AuthUserEntity(String username, String password, Collection<? extends GrantedAuthority> authorities, Long id) {
+    public AuthUserEntity(String username, String password, Collection<? extends GrantedAuthority> authorities, Integer id) {
         super(username, password, authorities);
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
